finished syncing, play verify

+++ app/src/main/java/com/browntowndev/liftlab/core/persistence/repositories/LiftMetricChartsRepository.kt
@@ -1,36 +1,45 @@
 package com.browntowndev.liftlab.core.persistence.repositories
 
 import androidx.compose.ui.util.fastMap
-import com.browntowndev.liftlab.core.common.FirebaseConstants
+import com.browntowndev.liftlab.core.common.FirestoreConstants
+import com.browntowndev.liftlab.core.common.fireAndForgetSync
 import com.browntowndev.liftlab.core.persistence.dao.LiftMetricChartsDao
 import com.browntowndev.liftlab.core.persistence.dtos.LiftMetricChartDto
 import com.browntowndev.liftlab.core.persistence.entities.LiftMetricChart
 import com.browntowndev.liftlab.core.persistence.entities.copyWithFirestoreMetadata
 import com.browntowndev.liftlab.core.persistence.mapping.FirebaseMappers.toEntity
-import com.browntowndev.liftlab.core.persistence.mapping.FirebaseMappers.toFirebaseDto
+import com.browntowndev.liftlab.core.persistence.mapping.FirebaseMappers.toFirestoreDto
 import com.browntowndev.liftlab.core.persistence.sync.FirestoreSyncManager
+import kotlinx.coroutines.CoroutineScope
 
 
 class LiftMetricChartsRepository(
     private val liftMetricChartsDao: LiftMetricChartsDao,
     private val firestoreSyncManager: FirestoreSyncManager,
+    private val syncScope: CoroutineScope,
 ): Repository {
     suspend fun deleteAllWithNoLifts() {
         val toDelete = liftMetricChartsDao.getAllWithNoLift()
+        if (toDelete.isEmpty()) return
+
         liftMetricChartsDao.deleteMany(toDelete)
-        firestoreSyncManager.deleteMany(
-            collectionName = FirebaseConstants.LIFT_METRIC_CHARTS_COLLECTION,
-            firestoreIds = toDelete.mapNotNull { it.firestoreId },
-        )
+
+        syncScope.fireAndForgetSync {
+            firestoreSyncManager.deleteMany(
+                collectionName = FirestoreConstants.LIFT_METRIC_CHARTS_COLLECTION,
+                firestoreIds = toDelete.mapNotNull { it.firestoreId },
+            )
+        }
     }
 
     suspend fun delete(id: Long) {
-        liftMetricChartsDao.get(id)?.let { toDelete ->
-            liftMetricChartsDao.delete(toDelete)
+        val toDelete = liftMetricChartsDao.get(id) ?: return
+        liftMetricChartsDao.delete(toDelete)
 
-            if (toDelete.firestoreId != null) {
+        if (toDelete.firestoreId != null) {
+            syncScope.fireAndForgetSync {
                 firestoreSyncManager.deleteSingle(
-                    collectionName = FirebaseConstants.LIFT_METRIC_CHARTS_COLLECTION,
+                    collectionName = FirestoreConstants.LIFT_METRIC_CHARTS_COLLECTION,
                     firestoreId = toDelete.firestoreId!!,
                 )
             }
@@ -54,16 +63,22 @@ class LiftMetricChartsRepository(
             )
         }
 
-        val insertedIds = liftMetricChartsDao.upsertMany(charts)
-        firestoreSyncManager.syncMany(
-            collectionName = FirebaseConstants.LIFT_METRIC_CHARTS_COLLECTION,
-            entities = charts.fastMap { it.toFirebaseDto() },
-            onSynced = { firebaseEntities ->
-                liftMetricChartsDao.updateMany(firebaseEntities.fastMap { it.toEntity() })
-            }
-        )
+        val upsertIds = liftMetricChartsDao.upsertMany(charts)
+        val chartsWithUpdatedIds = charts.zip(upsertIds).map { (chart, id) ->
+            if (id == -1L) chart else chart.copy(id = id)
+        }
 
-        return insertedIds
+        syncScope.fireAndForgetSync {
+            firestoreSyncManager.syncMany(
+                collectionName = FirestoreConstants.LIFT_METRIC_CHARTS_COLLECTION,
+                entities = chartsWithUpdatedIds.fastMap { it.toFirestoreDto() },
+                onSynced = { firebaseEntities ->
+                    liftMetricChartsDao.updateMany(firebaseEntities.fastMap { it.toEntity() })
+                }
+            )
+        }
+
+        return upsertIds
     }
 
     suspend fun upsert(liftMetricChart: LiftMetricChartDto): Long {
@@ -79,16 +94,21 @@ class LiftMetricChartsRepository(
                 synced = false,
             )
 
-        val insertId = liftMetricChartsDao.upsert(toUpsert)
-        firestoreSyncManager.syncSingle(
-            collectionName = FirebaseConstants.LIFT_METRIC_CHARTS_COLLECTION,
-            entity = toUpsert.toFirebaseDto(),
-            onSynced = { firebaseEntity ->
-                liftMetricChartsDao.update(firebaseEntity.toEntity())
-            }
-        )
+        val upsertId = liftMetricChartsDao.upsert(toUpsert).let {
+            if (it == -1L) toUpsert.id else it
+        }
 
-        return insertId
+        syncScope.fireAndForgetSync {
+            firestoreSyncManager.syncSingle(
+                collectionName = FirestoreConstants.LIFT_METRIC_CHARTS_COLLECTION,
+                entity = toUpsert.toFirestoreDto().copy(id = upsertId),
+                onSynced = { firebaseEntity ->
+                    liftMetricChartsDao.update(firebaseEntity.toEntity())
+                }
+            )
+        }
+
+        return upsertId
     }
 
     suspend fun getAll(): List<LiftMetricChartDto> {
